chore: move non cmd files to internal
